agent_adk/pyproject.toml
[project]
name = "adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.51.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.2.1",
]
